rename the DM2 file
