+++ src/components/diagnosticManager.ts
