+++ src/features/repairCode/buildRuleBasedRepairCodeJobs.ts
@@ -58,4 +58,4 @@ export const buildRuleBasedRepairCodeJobs = (
         lengths,
         version,
     );
-}   
+}
